+++ src/i18n/translate-i18n/index.ts
@@ -1,5 +1,5 @@
-import type { ApiOptions, Translator } from '../base/types';
-import translate from '../base/index';
+import type { ApiOptions, Translator } from '../../base/types';
+import translate from '../../base/index';
 import type { Code, I18nOptions } from './types';
 import antd from './languages/antd';
 export function getCode(code: Code, type: Translator) {
