+++ src/i18n/translate-i18n/languages/genTargetCode.ts
@@ -1,9 +1,9 @@
 // 谷歌云api规范
 //https://cloud.google.com/translate/docs/languages?hl=zh-cn
-import google from '../../base/languages/google';
+import google from '../../../base/languages/google';
 // 有道api规范
 // https://ai.youdao.com/DOCSIRMA/html/%E8%87%AA%E7%84%B6%E8%AF%AD%E8%A8%80%E7%BF%BB%E8%AF%91/API%E6%96%87%E6%A1%A3/%E6%96%87%E6%9C%AC%E7%BF%BB%E8%AF%91%E6%9C%8D%E5%8A%A1/%E6%96%87%E6%9C%AC%E7%BF%BB%E8%AF%91%E6%9C%8D%E5%8A%A1-API%E6%96%87%E6%A1%A3.html
-import youdao from '../../base/languages/youdao';
+import youdao from '../../../base/languages/youdao';
 // 和umi库的SelectLang对标 import {SelectLang} from 'umi';
 import antd from './source/antd';
 import { fsExtra } from '@txpjs/utils-node';
